pyproject.toml
[project]
name = "bigcord-tpl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "Pillow>=10.0.0",
    "requests>=2.28.0",
]
